+++ src/main/java/kr/hhplus/be/server/api/reservation/application/service/ReservationService.java
@@ -3,13 +3,11 @@ package kr.hhplus.be.server.api.reservation.application.service;
 import jakarta.transaction.Transactional;
 import kr.hhplus.be.server.api.common.exception.CustomException;
 import kr.hhplus.be.server.api.common.type.ReservationStatus;
-import kr.hhplus.be.server.api.common.type.SeatStatus;
-import kr.hhplus.be.server.api.concert.exception.SeatErrorCode;
 import kr.hhplus.be.server.api.reservation.application.dto.PaymentServiceRequest;
 import kr.hhplus.be.server.api.reservation.application.dto.ReservationServiceRequest;
 import kr.hhplus.be.server.api.reservation.domain.entity.Reservation;
 import kr.hhplus.be.server.api.concert.domain.entity.Seat;
-import kr.hhplus.be.server.api.reservation.domain.entity.repository.ReservationRepository;
+import kr.hhplus.be.server.api.reservation.domain.repository.ReservationRepository;
 import kr.hhplus.be.server.api.reservation.exception.ReservationErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
