+++ src/main/java/kr/hhplus/be/server/api/reservation/domain/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.api.reservation.domain.entity.repository;
+package kr.hhplus.be.server.api.reservation.domain.repository;
 
 import kr.hhplus.be.server.api.reservation.domain.entity.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
